Added spaceship, skybox, moving earth, fixed some blending and culling-errors

+++ ParticleSystem/HUD.cpp
@@ -23,7 +23,7 @@ HUD::~HUD()
     ;
 }
 
-void HUD::update(int delta)
+void HUD::update(double delta)
 {
     ;
 }
